add comments, clean code, organize merkle_partial exports
