+++ eth2/utils/merkle_partial/src/lib.rs
@@ -1,47 +1,19 @@
-pub mod cache;
-pub mod error;
+mod cache;
+mod error;
 pub mod field;
-pub mod merkle_tree_overlay;
-pub mod partial;
-pub mod path;
+mod merkle_tree_overlay;
+mod partial;
+mod path;
+mod ser;
 pub mod tree_arithmetic;
 
-use tree_hash::BYTES_PER_CHUNK;
+pub use cache::Cache;
+pub use error::Error;
+pub use merkle_tree_overlay::{impls, MerkleTreeOverlay};
+pub use partial::Partial;
+pub use path::Path;
+pub use ser::SerializedPartial;
 
 pub type NodeIndex = u64;
 
-/// A serializable represenation of a merkle proof
-#[derive(Clone, Debug, Default)]
-pub struct SerializedPartial {
-    pub indices: Vec<NodeIndex>,
-    pub chunks: Vec<u8>, // vec<bytes32>
-}
-
-impl PartialEq for SerializedPartial {
-    fn eq(&self, other: &Self) -> bool {
-        if self.indices.len() != other.indices.len() || self.chunks.len() != other.chunks.len() {
-            return false;
-        }
-
-        self.indices.iter().enumerate().fold(true, |s, (i, j)| {
-            let chunk = self.chunks[i * BYTES_PER_CHUNK];
-
-            if let Some(other_index) = other.indices.iter().position(|e| e == j) {
-                let other_chunk = other.chunks[other_index * BYTES_PER_CHUNK];
-                return s && chunk == other_chunk;
-            }
-
-            false
-        })
-    }
-}
-
-#[macro_export]
-macro_rules! vec_to_array {
-    ($vec:expr, $array_size:literal) => {{
-        let mut array = [0; $array_size];
-        let bytes = &$vec[..array.len()];
-        array.copy_from_slice(bytes);
-        array
-    }};
-}
+use tree_hash::BYTES_PER_CHUNK;
