+++ eth2/utils/merkle_partial_derive/Cargo.toml
@@ -1,7 +1,7 @@
 [package]
 name = "merkle_partial_derive"
-version = "0.1.0"
-authors = ["Matt Garnett <matt.garnett@consensys.net>"]
+version = "0.0.0"
+authors = ["Matt Garnett <matt.garnett@consensys.net>", "Paul Hauner <paul@sigmaprime.io>"]
 edition = "2018"
 description = "Procedural derive macros to accompany the merkle_partial crate."
 
@@ -14,4 +14,6 @@ syn = { version = "0.15", features = ["extra-traits"] }
 quote = "0.6"
 merkle_partial = { path = "../merkle_partial" }
 proc-macro2 = "0.4.30"
+
+[dev-dependencies]
 ethereum-types = "0.6.0"
